+++ core/server/api/index.js
@@ -26,7 +26,7 @@ var _ = require('lodash'),
     subscribers = require('./subscribers'),
     authentication = require('./authentication'),
     uploads = require('./upload'),
-    exporter = require('../data/export'),
+    exporter = require('../data/exporter'),
     slack = require('./slack'),
     webhooks = require('./webhooks'),
     oembed = require('./oembed'),
