+++ core/test/unit/data/exporter/index_spec.js
@@ -4,7 +4,7 @@ var should = require('should'),
     Promise = require('bluebird'),
     db = require('../../../../server/data/db'),
     common = require('../../../../server/lib/common'),
-    exporter = rewire('../../../../server/data/export'),
+    exporter = rewire('../../../../server/data/exporter'),
     schema = require('../../../../server/data/schema'),
     models = require('../../../../server/models'),
     schemaTables = Object.keys(schema.tables),
